% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_likert.R
\name{likert}
\alias{likert}
\title{Position Likert-type data}
\usage{
likert(nudge_y = 0, reverse = FALSE)
}
\arguments{
\item{reverse}{If TRUE, will reverse the default stacking order. This is
useful if you're rotating both the plot and legend.}

\item{nudge_x}{A horizontal adjustment useful for positing text labels next
to or on top of the bars rather than inside them.}
}
\description{
`position_likert()` positions bars to improve the legibility of Likert-type
data. This means it will stack bars on top of each other but center them at
the mid-point. The positioning works for both bar geoms and text labels.
}
