% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_likert.R
\name{position_likert}
\alias{position_likert}
\title{Position Likert-type data}
\usage{
position_likert()
}
\description{
\code{position_likert()} positions bars to improve the legibility of Likert-type
data. This means it will stack bars on top of each other but center them at
the mid-point. The positioning works for both bar geoms and text labels.
}
