% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_vhistogram.R
\name{stat_vhistogram}
\alias{stat_vhistogram}
\title{Plot vertical histograms per group}
\usage{
stat_vhistogram(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  bins = NULL,
  binwidth = NULL,
  center = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{x}{A character indicating whom to say hello to. 'World' by default.}
}
\description{
This stat plots the distribution of a variable for each group. It is
particularly suitable for visualizing distributions of categorical data such
as Likert-scale data.
}
