% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_vhistogram.R
\name{stat_vhistogram}
\alias{stat_vhistogram}
\title{Plot vertical histograms per group}
\usage{
stat_vhistogram(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  bins = NULL,
  binwidth = NULL,
  center = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by [aes()]. If specified and
`inherit.aes = TRUE` (the default), it is combined with the default mapping
at the top level of the plot. You must supply `mapping` if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If `NULL`, the default, the data is inherited from the plot
   data as specified in the call to [ggplot()].

   A `data.frame`, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   [fortify()] for which variables will be created.

   A `function` will be called with a single argument,
   the plot data. The return value must be a `data.frame`, and
   will be used as the layer data. A `function` can be created
   from a `formula` (e.g. `~ head(.x, 10)`).}

\item{geom}{The geometric object to use to display the data, either as a
`ggproto` `Geom` subclass or as a string naming the geom stripped of the
`geom_` prefix (e.g. `"point"` rather than `"geom_point"`)}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. `"jitter"` to use `position_jitter`), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{bins}{Number of bins.}

\item{binwidth}{The width of the bins.}

\item{center}{Whether the bins should be centered on their category.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with a
warning. If `TRUE``, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
`NA`, the default, includes if any aesthetics are mapped.
`FALSE` never includes, and `TRUE` always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. [borders()].}
}
\description{
This stat plots the distribution of a variable for each group. It is
particularly suitable for visualizing distributions of categorical data such
as Likert-scale data.
}
\details{
# Aesthetics

  `stat_vhistogram` understands the following (required) aesthetics:
  * x
  * y
}
\seealso{
`hist()`
}
